refactor(logic): split capital_and_gp_fall into part-of concern files (Refs #3290)
Phase 0 prerequisite for the colonizethis_logic package split (#3290):
decompose world/capital_and_gp_fall.dart (514 lines, > the 500-line Phase 0
target) into behavior-preserving `part`/`part of` concern files, mirroring
the established pattern (game_setup_create #3314, locked_province_assigner
#3316).

- capital_and_gp_fall.dart (root, 72): shared public surface only —
  CapitalReassignmentEligibility + evaluateCapitalReassignmentEligibility +
  part directives.
- capital_and_gp_fall_reassignment.dart (part): post-combat capital
  reassignment for GPs (applyCapitalReassignmentAfterCombat) and
  Minor/Tribe factions (applyFactionCapitalReassignmentAfterCombat) plus the
  shared _resolveReassignmentTileOrThrow helper.
- capital_and_gp_fall_terminal.dart (part): terminal fall —
  applyFactionTerminalFall + _applyTerminalFallForFaction, and
  applyGreatPowerFall.

Pure structural extraction (no algorithm or public API change); all moved
members keep their original (public/private) visibility within the single
library via part-of, so external importers and tests are unaffected. Removed
the file from tool/logic_source_file_size_baseline.json so the 500-line gate
now actively enforces it, and made the Phase 0 SPEC AC count-robust.

Refs #3290
